--- gillespy2/test.py
@@ -1,5 +1,0 @@
-def capital_case(x):
-    return x.capitalize()
-
-def test_capital_case():
-    assert capital_case('semaphore') == 'Semaphore'
